CVBot/requirements.txt
pandas==1.5.0
matplotlib==3.6.0
python-telegram-bot==20.0
